+++ demo/src/lib.rs
@@ -4,10 +4,11 @@ mod plugins;
 mod windows;
 
 use std::collections::HashMap;
+use std::{cell::Cell, rc::Rc};
 
 use crate::plugins::ImagesPluginData;
 use egui::Context;
-use walkers::{Map, MapMemory, Tiles, TilesManager};
+use walkers::{Map, MapMemory, Position, Tiles, TilesManager};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum Provider {
@@ -80,6 +81,7 @@ pub struct MyApp {
     selected_provider: Provider,
     map_memory: MapMemory,
     images_plugin_data: ImagesPluginData,
+    last_click_position: Rc<Cell<Option<Position>>>,
 }
 
 impl MyApp {
@@ -94,6 +96,7 @@ impl MyApp {
             selected_provider: Provider::OpenStreetMap,
             map_memory: MapMemory::default(),
             images_plugin_data,
+            last_click_position: Rc::new(Cell::new(None)),
         }
     }
 }
@@ -109,7 +112,10 @@ impl eframe::App for MyApp {
             .frame(rimless)
             .show(ctx, |ui| {
                 // Typically this would be a GPS acquired position which is tracked by the map.
-                let my_position = places::wroclaw_glowny();
+                let my_position = match self.last_click_position.get() {
+                    Some(position) => position,
+                    None => places::wroclaw_glowny(),
+                };
 
                 let tiles = self
                     .providers
@@ -125,7 +131,10 @@ impl eframe::App for MyApp {
                 let map = map
                     .with_plugin(plugins::places())
                     .with_plugin(plugins::images(&mut self.images_plugin_data))
-                    .with_plugin(plugins::CustomShapes {});
+                    .with_plugin(plugins::CustomShapes {})
+                    .with_plugin(plugins::ClickWatcher {
+                        last_click: self.last_click_position.clone(),
+                    });
 
                 // Draw the map widget.
                 ui.add(map);
@@ -136,6 +145,7 @@ impl eframe::App for MyApp {
 
                     zoom(ui, &mut self.map_memory);
                     go_to_my_position(ui, &mut self.map_memory);
+                    show_my_position(ui, &my_position);
                     controls(
                         ui,
                         &mut self.selected_provider,